+++ Changelog.md
@@ -7,6 +7,7 @@
   - Generator HW:
     * Support generics in field length and array depth: field length/depth is decided by max value, but MSBs above the parameter value are forced to 0 (suppose synthesis will properly simplify it)
     * New option "rifmux_pipe_invalid" to add a pipe level to the generation of invalid address flag when no RIF is selected. Allows to relax timing constraint on this signal.
+    * Update APB bridge to remove unnecessary wait cycle for write access
   - Generator RAL: new option "--ral_force_base" allows to force base class of derived register to still use uvm_reg
 
 ## [0.22.3] - 2026-05-08